extract surge utils code
